pyproject.toml
[project]
name = "genai-with-python-3-0"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-qdrant>=1.1.0",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "playsound==1.2.2",
    "pydantic>=2.12.5",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
